shoezone_backend/shoeZone/requirements.txt
django==4.0.3
mysqlclient==2.1.0
djangorestframework==3.13.1
django-cors-headers==3.11.0